requirements.txt
# Raspberry Pi 5 (64-bit Bookworm) için
picamera2>=0.3.12
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pyttsx3>=2.90
gTTS>=2.3.0
pygame>=2.5.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
requests>=2.31.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
google-generativeai>=0.8.0
Pillow>=10.0.0
scipy>=1.11.0
plyfile>=1.0.0
